Done: 1. Fix unhandling barrier flags in EXT4. 2. Fix missing barrier problem in ext4 delay allocation write. 3. Reason of freezing is discovered. :wait_on_dispatch() does not be awaken in some condition especially when running benchmark. Now I didn't found the proper reason of this situation. :However, this infinite waiting can be resolved by waking up it in I/O complete. :This may degrade the performance, but it rarely occurs.
Todo:
Finding out precise reason of missing wakeup signal.

+++ include/linux/fs.h
@@ -2421,7 +2421,7 @@ static inline void remove_inode_hash(struct inode *inode)
 extern void inode_sb_list_add(struct inode *inode);
 
 #ifdef CONFIG_BLOCK
-extern void submit_bio(int, struct bio *);
+extern void submit_bio(long long/*int*/, struct bio *);
 /*UFS*/
 extern void submit_bio64(long long, struct bio *);
